Add quantity selector
Also fixed some bugs with the size selector. In order for the size
selector to be updated when the styles changed we need to useEffect
suscribed to the skus prop. However, in the future we will want to
change the initial selectedSize to "Select Size" which will remove the
need for that subscription.

In principle quantities should not be state since it can be computed
from the selectedSize. However, in order to get the <Select /> for
quantity to show the correct options I added it to state to trigger a
rerender that would include the updated value for quantities. Otherwise,
the <Select /> is empty, maybe due to the async nature of useEffect.