Add AddToCart button
One of the specifications for the button requires that if no size is
selected, clicking the Add to Cart button should open the size dropdown.
This might be very difficult to implement:

https://stackoverflow.com/questions/249192/
https://medium.com/browserquirks/
  browserquirk-programmatically-opening-a-select-box-4ca745a8468f

Note that changes to Chrome that prevent untrusted events from executing
the default action (i.e. events from dispatchEvent), which aligned it
with Firefox and IE, mean that old proposed solutions to this are not
valid.

https://bugs.chromium.org/p/chromium/issues/detail?id=520520

Furthermore, triggering a click event with selectRef.click() will
only foucs the select element and will NOT open the dropdown.

https://stackoverflow.com/questions/32653002/

There are CSS hacks to simulate opening of the dropdown by modifying the
size property onFocus and resetting it onBlur:

https://stackoverflow.com/questions/9870174/
https://stackoverflow.com/questions/360431/

There are also implications for testing and trying to open a select:
https://github.com/ant-design/ant-design/issues/23044