Image gallery basic functionality (#95)
* Add prop-types package

* Create basic carousel component

* Remove unnecessary range function

* Clean up App

* Remove empty Carousel test file

* Set up data flow in Overview

* Add product details to ProductInformation

* Make ProductInformation props required

* Fix props for StyleSelector

* Use proper async useEffect pattern

* Select style by ID

Optional chaining is used to avoid errors resulting from lookups on
undefined.

* Add basic style selector functionality

* Add size selector

* Add quantity selector

Also fixed some bugs with the size selector. In order for the size
selector to be updated when the styles changed we need to useEffect
suscribed to the skus prop. However, in the future we will want to
change the initial selectedSize to "Select Size" which will remove the
need for that subscription.

In principle quantities should not be state since it can be computed
from the selectedSize. However, in order to get the <Select /> for
quantity to show the correct options I added it to state to trigger a
rerender that would include the updated value for quantities. Otherwise,
the <Select /> is empty, maybe due to the async nature of useEffect.

* Filter SKUs by availability

Fix needed: currently not possible to select a size in <Select />

* Change subscription for initial size

This effect was being called repeatedly, preventing the size <Select />
from displaying the correct value since it was being overwritten.

Also remove use of the default keyword as a paramter for <Select />

* Add AddToCart button

One of the specifications for the button requires that if no size is
selected, clicking the Add to Cart button should open the size dropdown.
This might be very difficult to implement:

https://stackoverflow.com/questions/249192/
https://medium.com/browserquirks/
  browserquirk-programmatically-opening-a-select-box-4ca745a8468f

Note that changes to Chrome that prevent untrusted events from executing
the default action (i.e. events from dispatchEvent), which aligned it
with Firefox and IE, mean that old proposed solutions to this are not
valid.

https://bugs.chromium.org/p/chromium/issues/detail?id=520520

Furthermore, triggering a click event with selectRef.click() will
only foucs the select element and will NOT open the dropdown.

https://stackoverflow.com/questions/32653002/

There are CSS hacks to simulate opening of the dropdown by modifying the
size property onFocus and resetting it onBlur:

https://stackoverflow.com/questions/9870174/
https://stackoverflow.com/questions/360431/

There are also implications for testing and trying to open a select:
https://github.com/ant-design/ant-design/issues/23044

* Add validation for AddToCart

Also removes useRef since the aim had been to useRef to focus or click
the select and open it programatically, but it is not possible to open a
select in this way.

* Add product details to ProductInformation

* Select style by ID

Optional chaining is used to avoid errors resulting from lookups on
undefined.

* Add basic style selector functionality

* Remove merge artifacts

* Test for style change on thumbnail click

* Resolve some style errors

* Add tests TODOs for AddToCart

* Move StyleSelector and AddToCart to Overview

* Check for selectedSize instead of selectedQuantity

* Add label for Select

* Reorganize mock data in Overview

* Test for correct sizes in selector

* Add test TODOs for ImageGallery

* Add basic image gallery functionality

* Add prop validation for ImageGallery

* Fix syntax error

* Add basic styling for carousel

* Test for first image shown by default

* Properly situate key for photos

* Fix off by 1 error in carousel

* Add photo_ids to photos