fix(Stars): fix hover functionality (#113)
* fix(Stars): fix hover functionality

The prop 'value' now takes a number and all internal manipulations of
values within Stars expect a number. Previously values would be coerced
to strings during their assignment to the prop 'value' of their button
and this string would be passed to the parent through
event.currentTarget.value. Setting event.currentTarget.value to 0 to
unset an interactive Stars instance overwrites the button's value (since
that is what event.currentTarget.value points to), so the API of Stars
was changed to allow Stars to report to its parent component what the
rating should now be.

* fix: revert non-PR files to main