feat: update styling of Overview components (#99)
* Make photo_id a string

* Move Carousel to shared directory

* Move global styles to globalStyles.js

* Add space and text size design tokens

Uses CSS variables following this approach:
https://www.joshwcomeau.com/css/css-variables-for-react-devs/

* Center App

* Enable data fetching

* Add Navbar

* Add fonts

* Refactor Carousel to hide items

* Add light/dark mode

* Implement Button and ImageButton

* Add SVG arrow icons to Carousel

* feat(overview): select main image by clicking thumbnail

* refactor(carousel): switch to controlled component

This facilitates selecting a main image by clicking on the thumbnail
(#43) and scrolling the thumbnail gallery if the thumbnail for the main
image is not in view (#45).

* refactor(carousel): Simplify scroll handling

* refactor(carousel): rename scrollTo -> scrollIndex

* feat(productInfo): add basic styling

* refactor(select): Move select to shared directory

Also implement basic styling for ProductInformation.

* feat(button): add primary button variant

* feat(select): add disabling and defaults

* refactor(select): rename defaultSelection -> selectionPrompt

* refactor(AddToCart): rename boolean

* build: add better styled-components debugging

Generates more readable class names during development.

* feat: add global styles and Navbar

* feat: add Button and ImageButton

* feat: add global styles and Navbar

* docs(shared): Add docs for Button/ImageButtons

* feat(shared): Add Select component

* docs(shared): Add docs for Select

* test(Select): Test for correct options list

Remove defaultValue prop since the expectation is that Select will be
used as a controlled component and the value prop and defaultValue prop
should not be used together. The value prop is for controlled components
and the defaultValue prop is for uncontrolled components.

* ci: generate coverage badge

* ci: change event trigger

* ci: fix indentation

* ci: update script version

* ci: fix coverage script

* ci: update README badge link

* ci: fix coverage script

* ci: change jest logo color

* ci: change jest logo color

* ci: fix badge link

* ci: simplify coverage script

* ci: fix coverage script

* ci: fix coverage script

* ci: ensure tests produce coverage

* ci: run coverage on push to main

* feat(StyleSelector): add border to selected

* feat(stars): Add several variations of stars

* feat(stars): make star show fractional value

* fix(stars): create StarIconFills to hold gradients

If the linear gradient is defined in StarIcon instead of StarIconFills
(which needs to appear in the DOM), all the StarIcon instances would
have the same fill value since all the linear gradients would have the
same id. # Please enter the commit message for your changes. Lines
starting

* feat(Stars): Add Stars components

Stars can be used to display a product rating statically or be used
interactively as a form to set a rating as in a review.

* feat(stars): show stars on product info

* docs(stars): docs for StarIcon and fills

* refactor(Stars): simplify Stars

* docs(Stars): add docs for Stars

* fix(Stars): don't use buttons when static

* feat(Stars): make Stars more accesible

* test(Stars): test correct value reporting

* on main with only stars changes

* fix(Icon): fix syntax error

* fix: restore App/Carousel

* fix(ArrowIcon): change to sharp chevron

* feat(Carousel): add ability to offset arrows

* refactor(Carousel): use transform to shift arrows

* Revert "refactor(Carousel): use transform to shift arrows"

This reverts commit 8143b707779132bfb5311540d644f28cfbb945d3.

Using 'transform: translate()' instead of changing the margin of the
arrows artificially increases the width of the carousel. It leaves the
space originally occupied by the arrow buttons as part of the carousel's
box.

* feat(Carousel): allow tweaks to button styling

* docs(Carousel): add docs for styling props

* feat(Carousel): add more styling control

* fix(Icon): allow for automatic scaling

If only one of iconHeight or iconWidth is provided, the assumption is
that Icon should be scaled to match that the given dimension.

* feat(Carousel): finalize Carousel functionality

* fix: update components to new Carousel

* docs(Carousel): add docs for styling props

* fix: revert non-PR files to main

* fix(Stars): fix hover functionality

The prop 'value' now takes a number and all internal manipulations of
values within Stars expect a number. Previously values would be coerced
to strings during their assignment to the prop 'value' of their button
and this string would be passed to the parent through
event.currentTarget.value. Setting event.currentTarget.value to 0 to
unset an interactive Stars instance overwrites the button's value (since
that is what event.currentTarget.value points to), so the API of Stars
was changed to allow Stars to report to its parent component what the
rating should now be.

* test(ImageGallery): use role to query

* test(ProductInformation): search for style by text

* style(Overview): move styling to bottom

* test(AddToCart): find Select by role

* test(StyleSelector): find style button by role

* fix(Stars): use keys for static StarIcons